Figured out dispose, cleaned and optimized some stuff.

+++ src/game/ecs/components/target_position.rs
@@ -4,6 +4,6 @@ use specs_derive::Component;
 #[derive(Default, Component, Debug)]
 #[storage(VecStorage)]
 pub struct TargetPosition {
-    pub x: u32,
-    pub y: u32,
+    pub x: usize,
+    pub y: usize,
 }
